destination object now has a category name in it when returned

+++ backend/src/main/java/com/techelevator/controller/DestinationController.java
@@ -30,7 +30,7 @@ public class DestinationController {
 	}
 	
 	@PostMapping("/profile/{username}/{destinationId}") 
-	public void checkIn(@PathVariable String username, @PathVariable Integer destinationId) {
+	public void checkIn(@PathVariable String username, @PathVariable Long destinationId) {
 		userDao.checkIntoDestination(username, destinationId);
 	}
 }
